backend/requirements.txt
Flask
flask-cors
gtts
Pillow
pydub
google-genai
python-dotenv
